commit TelaDePatrimonio

+++ Hermes/src/controller/TabelaController.java
@@ -34,17 +34,21 @@ public class TabelaController  extends ComponentAdapter {
 	}
 
 	private void excluirPatrimonio() {
-		int linhaSelecionada = telaTabelaPatrimonios.getTable().getSelectedRow();
-
-		String idPatrimonio = (String) telaTabelaPatrimonios.getTable().getValueAt(linhaSelecionada, 0);
-
-		int confirm = JOptionPane.showConfirmDialog(null, "Deseja excluir?", "Confirmação", JOptionPane.YES_NO_OPTION);
-
-		if (confirm == JOptionPane.YES_OPTION) {
-			patrimonioDAO.excluirPatrimonio(idPatrimonio);
-
-		}
-
+	    int linha = telaTabelaPatrimonios.getLinhaSelecionada();
+	    if (linha == -1) {
+	        JOptionPane.showMessageDialog(null, "Selecione um patrimonio para excluir!");
+	        return;
+	    }
+	    
+	    int confirm = JOptionPane.showConfirmDialog(null, "Deseja excluir?", "Confirmação", JOptionPane.YES_NO_OPTION);
+	    if (confirm == JOptionPane.YES_OPTION) {
+	    	PatrimonioDAO model = (PatrimonioDAO) telaTabelaPatrimonios.getTable().getModel();
+	    	int idPatrimonio = 0;
+	 
+	        patrimonioDAO.excluirPatrimonio(idPatrimonio); 
+	        telaTabelaPatrimonios.atualizarTabela();
+	        JOptionPane.showMessageDialog(null, "Produto excluído com sucesso!");
+	    }
 	}
 
 	public void componentShown(ComponentEvent e) {
